Updated wallpaper with actual BMP Image

+++ src/bootloader/stage2/memory.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "stdint.h"
 
-void* memcpy(void* dst, const void* src, uint16_t num);
-void* memset(void* ptr, int value, uint16_t num);
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
+void* memcpy(void* dst, const void* src, uint32_t num);
+void* memset(void* ptr, int value, uint32_t num);
+int memcmp(const void* ptr1, const void* ptr2, uint32_t num);
